test: preserve WhatsApp max-lines ratchet path
